feat: 전체, 내 제휴 토글 파일 별도 생성
